- making oscimap default backend - added about-screen - added TreeTile for Tile lookup, dropping that HashMap - using simple line-shader instead of std-derivatives one,   about twice as faster here - use distance calculation from MapRenderer - removing TileScheduler - no need for MapGeneratorJob, pass MapTile directly to MapWorkers - added two-finger tap gestures for zoom-in/out - added tub/tron rendertheme - started caching for oscimap - add x/y coordinates to MapPosition, using it in MapRenderer - create tag hash when needed - no need for long tile coordinates max zoomlevel 31 should suffice

+++ src/org/mapsforge/android/swrenderer/GLMapTile.java
@@ -36,7 +36,7 @@ public class GLMapTile extends MapTile {
 	 * @param zoomLevel
 	 *            ..
 	 */
-	public GLMapTile(long tileX, long tileY, byte zoomLevel) {
+	public GLMapTile(int tileX, int tileY, byte zoomLevel) {
 		super(tileX, tileY, zoomLevel);
 		mScale = 1;
 		mTextureID = -1;
